Initial commit
I am passionate about birdwatching. I often travel to new places in a bid to see bird species that I have not seen before. In preparation for these trips I usually assess on the potential "lifers" that I may get if I were to visit. For this I browse the information on the bird sightings on the hotspots in the eBird website. This program is to help extract the information and then compare with my "Life"list programmatically.